[REFACTOR] replaced osg::Vec2i with osg::Vec2f

+++ include/SpecialMatrix/FileParser.h
@@ -79,7 +79,7 @@ public:
 	 * @param int column
 	 * @return NodePos a new position of node
 	 */
-	osg::Vec2i getAvailablePosition( SpecialMatrix::NodeConnections* connections, int row, int column );
+	osg::Vec2f getAvailablePosition( SpecialMatrix::NodeConnections* connections, float row, float column );
 
 	/**
 	 * @brief createSettings initializes the used types
